Add credential types and persistent data storage

+++ docker/ghcr/entrypoint.sh
@@ -6,12 +6,21 @@ mkdir -p /data/system /data/audit
 for source in /opt/uhppoted/defaults/system/*.json; do
   target="/data/system/$(basename "$source")"
   if [ ! -e "$target" ]; then
-    cp "$source" "$target"
+    legacy="/usr/local/etc/uhppoted/httpd/system/$(basename "$source")"
+    if [ -e "$legacy" ]; then
+      cp "$legacy" "$target"
+    else
+      cp "$source" "$target"
+    fi
   fi
 done
 
 if [ ! -e /data/auth.json ]; then
-  cp /opt/uhppoted/defaults/auth.json /data/auth.json
+  if [ -e /usr/local/etc/uhppoted/httpd/auth.json ]; then
+    cp /usr/local/etc/uhppoted/httpd/auth.json /data/auth.json
+  else
+    cp /opt/uhppoted/defaults/auth.json /data/auth.json
+  fi
 fi
 
 exec /opt/uhppoted/uhppoted-httpd --debug --config /usr/local/etc/uhppoted/uhppoted.conf --console
